+++ src/components/content/medications/patient-medications.stories.tsx
@@ -1,21 +1,21 @@
 import {
-  ProviderMedsTable,
-  ProviderMedsTableProps,
-} from "@/components/content/medications/provider-meds-table";
+  PatientMedications,
+  PatientMedicationsProps,
+} from "@/components/content/medications/patient-medications";
 import { CTWProvider } from "@/components/core/ctw-provider";
 import { PatientProvider } from "@/components/core/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
 import type { Meta, StoryObj } from "@storybook/react";
 import { setupMedicationMocks } from "./story-helpers/mocks/requests";
 
-type Props = ProviderMedsTableProps;
+type Props = PatientMedicationsProps;
 
 export default {
   tags: ["docsPage"],
-  component: ProviderMedsTable,
+  component: PatientMedications,
   decorators: [
     (Story, { args }) => (
-      <CTWProvider env="dev" authToken="12345" builderId="12345">
+      <CTWProvider env="dev" authToken="ey.12345" builderId="12345">
         <PatientProvider patientID="007" systemURL={SYSTEM_ZUS_UNIVERSAL_ID}>
           <Story args={args} />
         </PatientProvider>
@@ -25,9 +25,4 @@ export default {
   ...setupMedicationMocks(),
 } as Meta<Props>;
 
-export const Basic: StoryObj<Props> = {
-  args: {
-    sortColumn: "display",
-    sortOrder: "asc",
-  },
-};
+export const Basic: StoryObj<Props> = {};